Refactor Laravel specific MR helpers
